+++ modulo3/projeto-labex/projeto-labex/README.md
@@ -1,11 +1,22 @@
 ### Projeto LabeX
 
 ### Descrição
-Página que simula um gerenciamento de viagens espaciais. Projeto desenvolvido durante o curso da Labenu_.
-Tecnologia utilizada: React. Bibliotecas utilizadas: axios, react-router, styled-components.
-A estilização do projeto foi realizada utilizando o Evergreen design system.
-Principais componentes: roterização das páginas, integração com API, autenticação e validação de formulários.
+- Página que simula um gerenciamento de viagens espaciais. Projeto desenvolvido durante o curso da Labenu_.
+- Tecnologia utilizada: React. Bibliotecas utilizadas: axios, react-router, styled-components.
+- Principais componentes: roteirização das páginas, integração com API, autenticação; validação de formulários; layout responsivo.
+- A estilização do projeto foi realizada utilizando o Evergreen Design System e a responsividade com a lib styled-components.
+- Login de acesso a área restrita: rodolfo@joganada.com.br / Senha: 4242
 
 ### Surge
-
 https://jrvc-labex.surge.sh/
+
+### Imagens
+Desktop:
+![image](https://user-images.githubusercontent.com/80327029/145694932-6b2146f5-e428-49eb-851c-20c06b86f6ca.png) 
+![image](https://user-images.githubusercontent.com/80327029/145694938-619893d8-30e2-44f1-bba7-99bd5efaa29c.png)
+![image](https://user-images.githubusercontent.com/80327029/145694953-4f50577a-f0c8-466a-8eaa-c51184147795.png) 
+![image](https://user-images.githubusercontent.com/80327029/145694959-38ea2ec0-4a61-4b6a-9186-6ac55de0e00a.png)
+![image](https://user-images.githubusercontent.com/80327029/145694971-8cd83852-5a8a-4aaf-9485-c52b0c08472c.png)
+
+Mobile:
+![image](https://user-images.githubusercontent.com/80327029/145727878-229838f0-0f26-40f6-b281-8475b9a44f92.png) ![image](https://user-images.githubusercontent.com/80327029/145727893-76a652da-b07b-497b-97ae-96f80e639311.png) ![image](https://user-images.githubusercontent.com/80327029/145727904-573f0636-c30b-4889-94e4-a068dc789e18.png) ![image](https://user-images.githubusercontent.com/80327029/145727933-1c3f99d8-cd43-4954-b585-83a547877888.png)
