+++ modulo3/projeto-labex/projeto-labex/src/components/ApplicationForm/styled-components.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const ContainerMobile = styled.div`  
+     @media (max-width: 600px) {
+         width: 100vw; 
+
+         form {
+             width: 80vw;
+         }
+        h2 {
+            font-size: 15px;
+        }
+        label {
+            font-size: 13px;
+        }
+    }
+`
