+++ modulo3/projeto-labex/projeto-labex/src/pages/AdminHomePage/AdminHomePage.js
@@ -1,10 +1,11 @@
-import { useProtectPage } from "../hooks/useProtectPage";
-import { useGetTrips } from "../hooks/useGetTrips";
+import { useProtectPage } from "../../hooks/useProtectPage";
+import { useGetTrips } from "../../hooks/useGetTrips";
 import { Pane, Spinner, Text, Table, Heading, IconButton, EditIcon, TrashIcon, majorScale, Button } from "evergreen-ui";
 import { useState } from "react";
 import { useHistory } from "react-router-dom";
-import CreateTrip from "../components/CreateTrip";
-import DeleteTrip from "../components/DeleteTrip";
+import CreateTrip from "../../components/CreateTrip/CreateTrip";
+import DeleteTrip from "../../components/DeleteTrip/DeleteTrip";
+import { ContainerMobile} from "./styled-components";
 
 const AdminHomePage = () => {
     const history = useHistory("");
@@ -15,11 +16,13 @@ const AdminHomePage = () => {
 
     useProtectPage();
 
+    // Função de logout e redireceiona para página de início:
     const logout = () => {
         localStorage.removeItem("token");
         history.push(`/`);
     }
 
+    // Funções para abertura de popups e páginas auxiliares:
     const abrirDetalhes = (id) => {
         history.push(`/admin/trips/${id}`);
     }
@@ -32,15 +35,16 @@ const AdminHomePage = () => {
         setDeletarViagem({ status: infoStatus, info: id });
     }
 
+    // Renderiza a lista de viagens na tela:
     const listaTrips = trips && trips.filter((elemento) => {
         return elemento.name.toUpperCase().includes(busca.toUpperCase());
     }).map((elemento) => {
         return (
             <Table.Row key={elemento.id} isSelectable>
-                <Table.TextCell flex={3}>{elemento.name}</Table.TextCell>
-                <Table.TextCell>
-                    <IconButton icon={EditIcon} marginRight={majorScale(2)} onClick={() => abrirDetalhes(elemento.id)} />
-                    <IconButton icon={TrashIcon} intent="danger" marginLeft={majorScale(2)} onClick={() => abrirDeletar(true, elemento)} />
+                <Table.TextCell flexBasis={0} flexShrink={0} flexGrow={4}>{elemento.name}</Table.TextCell>
+                <Table.TextCell minWidth="100px">
+                    <IconButton icon={EditIcon} marginRight={majorScale(1)} onClick={() => abrirDetalhes(elemento.id)} />
+                    <IconButton icon={TrashIcon} intent="danger" marginLeft={1} onClick={() => abrirDeletar(true, elemento)} />
                 </Table.TextCell>
             </Table.Row>
         )
@@ -55,7 +59,7 @@ const AdminHomePage = () => {
                 </Pane>
 
                 <Pane marginTop={10} display="flex" flexDirection="column" alignItems="center">
-                    <Pane width="50vw" display="flex" padding={10} justifyContent="center" alignItems="center" marginBottom={30}>
+                    <Pane display="flex" padding={10} justifyContent="center" alignItems="center" marginBottom={30}>
 
                         <Button intent="success" marginRight={majorScale(2)} onClick={() => abrirCriar(true)}>Criar viagem</Button>
                         <Button marginRight={majorScale(2)} onClick={logout}>Logout</Button>
@@ -65,14 +69,16 @@ const AdminHomePage = () => {
                     {!loadingTrips && trips && trips.length === 0 && <Text>Sem trips adicionadas.</Text>}
                     {!loadingTrips && errorTrips && <Text>Ocorreu um erro inesperado! Reinicie a página.</Text>}
                     {!loadingTrips && trips && trips.length > 0 && listaTrips &&
-                        <Table width="50vw">
-                            <Table.Head>
-                                <Table.SearchHeaderCell onChange={(value) => setBusca(value)} placeholder="Busca" flex={3} />
-                            </Table.Head>
-                            <Table.VirtualBody height={48 * trips.length}>
-                                {listaTrips}
-                            </Table.VirtualBody>
-                        </Table>
+                        <ContainerMobile>
+                            <Table width="100%">
+                                <Table.Head>
+                                    <Table.SearchHeaderCell onChange={(value) => setBusca(value)} placeholder="Busca" />
+                                </Table.Head>
+                                <Table.VirtualBody height={48 * trips.length}>
+                                    {listaTrips}
+                                </Table.VirtualBody>
+                            </Table>
+                        </ContainerMobile>
                     }
                 </Pane>
             </Pane>
