+++ modulo3/projeto-labex/projeto-labex/src/pages/AdminHomePage/styled-components.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const ContainerMobile = styled.div`
+        width: 40vw;
+
+     @media (max-width: 1075px) {
+        width: 90vw;
+}
+`
