+++ modulo3/projeto-labex/projeto-labex/src/pages/ListTripsPage/styled-components.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const ContainerMobile = styled.div`
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+     @media (max-width: 805px) {
+        div {
+     display: none;
+    }
+}
+`
